requirements.txt
discord.py
flask
pymongo
dnspython